fix: start script, minor errors

+++ main.go
@@ -23,8 +23,6 @@ func main() {
 		fmt.Printf("Error getting block storage info: %v", err)
 	}
 
-	fmt.Printf("%v\n", block)
-
 	for _, disk := range block.Disks {
 		litter.Dump(disk)
 		if len(disk.Partitions) == 0 && strings.Contains(disk.Name, "nvme") {
@@ -36,6 +34,12 @@ func main() {
 					os.Exit(1)
 				}
 			}
+			out, err := exec.Command("blkid", "-o", "value", "-s", "TYPE", devicePath).CombinedOutput()
+			if err == nil {
+				log.Printf("skipping making fs due to existing fs: output: %s\n", out)
+				continue
+			}
+			log.Printf("blkid fs type found: %s\n, error: %v", out, err)
 			if out, err := exec.Command("mkfs", "-t", "xfs", "-f", devicePath).CombinedOutput(); err != nil {
 				log.Printf("error making fs: output: %s\n, error: %v", out, err)
 				os.Exit(1)
@@ -43,6 +47,8 @@ func main() {
 			if err := mounter.Mount(devicePath, mountPath, "xfs", []string{}); err != nil {
 				log.Printf("error making mount: %v", err)
 			}
+		} else {
+			log.Printf("skipping non-nvme disk")
 		}
 	}
 }
